requirements.txt
arcade~=2.6.17
azure-cognitiveservices-speech==1.30.0
python-dotenv==1.0.0